+++ src/main/java/com/emall/entity/OrderItem.java
@@ -15,21 +15,21 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Repository
 public class OrderItem {
-    private Integer oiId;           //订单明细id
+    private Long oiId;              //订单明细id
 
-    private Integer oId;            //订单id
+    private Long oId;              //订单id
 
-    private Integer uId;            //订单用户id
+    private Long uId;              //订单用户id
 
-    private Integer gId;            //订单商品id
+    private Long gId;              //订单商品id
 
-    private String gName;           //订单商品名称
+    private String gName;         //订单商品名称
 
-    private String gImage;          //订单商品图片地址
+    private String gImage;         //订单商品图片地址
 
-    private BigDecimal gPrice;      //订单商品价格
+    private BigDecimal gPrice;     //订单商品价格
 
-    private String gCount;          //订单商品数量
+    private String gCount;        //订单商品数量
 
-    private BigDecimal oiSubtotal;  //订单明细小计
+    private BigDecimal oiSubtotal; //订单明细小计
 }
